fix: Enfore probabilistic proof late reveal
