+++ pkg/relayer/session/proof.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/pokt-network/poktroll/pkg/client"
 	"github.com/pokt-network/poktroll/pkg/crypto/protocol"
+	poktrand "github.com/pokt-network/poktroll/pkg/crypto/rand"
 	"github.com/pokt-network/poktroll/pkg/either"
 	"github.com/pokt-network/poktroll/pkg/observable"
 	"github.com/pokt-network/poktroll/pkg/observable/channel"
@@ -13,6 +14,7 @@ import (
 	"github.com/pokt-network/poktroll/pkg/observable/logging"
 	"github.com/pokt-network/poktroll/pkg/relayer"
 	"github.com/pokt-network/poktroll/x/proof/types"
+	prooftypes "github.com/pokt-network/poktroll/x/proof/types"
 	"github.com/pokt-network/poktroll/x/shared"
 )
 
@@ -222,17 +224,38 @@ func (rs *relayerSessionsManager) newMapProveSessionsFn(
 func (rs *relayerSessionsManager) goProveClaims(
 	ctx context.Context,
 	sessionTrees []relayer.SessionTree,
-	sessionPathBlock client.Block,
+	proofPathSeedBlock client.Block,
 	proofsGeneratedCh chan<- []relayer.SessionTree,
-	failSubmitProofsSessionsCh chan<- []relayer.SessionTree,
+	failGenerateProofsSessionsCh chan<- []relayer.SessionTree,
 ) {
 	logger := rs.logger.With("method", "goProveClaims")
+
+	// sessionTreesWithProofRequired will accumulate all the sessionTrees that
+	// will require a proof to be submitted.
+	sessionTreesWithProofRequired := make([]relayer.SessionTree, 0)
+	for _, sessionTree := range sessionTrees {
+		isProofRequired, err := rs.isProofRequired(ctx, sessionTree, proofPathSeedBlock)
+
+		// If an error is encountered while determining if a proof is required,
+		// do not create the claim since the proof requirement is unknown.
+		// Creating a claim and not submitting a proof (if necessary) could lead to a stake burn!!
+		if err != nil {
+			failGenerateProofsSessionsCh <- sessionTrees
+			rs.logger.Error().Err(err).Msg("failed to determine if proof is required, skipping claim creation")
+			continue
+		}
+
+		// If a proof is required, add the session to the list of sessions that require a proof.
+		if isProofRequired {
+			sessionTreesWithProofRequired = append(sessionTreesWithProofRequired, sessionTree)
+		}
+	}
 
 	// Separate the sessionTrees into those that failed to generate a proof
 	// and those that succeeded, then send them on their respective channels.
 	failedProofs := []relayer.SessionTree{}
 	successProofs := []relayer.SessionTree{}
-	for _, sessionTree := range sessionTrees {
+	for _, sessionTree := range sessionTreesWithProofRequired {
 		select {
 		case <-ctx.Done():
 			return
@@ -241,7 +264,7 @@ func (rs *relayerSessionsManager) goProveClaims(
 		// Generate the proof path for the sessionTree using the previously committed
 		// sessionPathBlock hash.
 		path := protocol.GetPathForProof(
-			sessionPathBlock.Hash(),
+			proofPathSeedBlock.Hash(),
 			sessionTree.GetSessionHeader().GetSessionId(),
 		)
 
@@ -258,6 +281,94 @@ func (rs *relayerSessionsManager) goProveClaims(
 		successProofs = append(successProofs, sessionTree)
 	}
 
-	failSubmitProofsSessionsCh <- failedProofs
+	failGenerateProofsSessionsCh <- failedProofs
 	proofsGeneratedCh <- successProofs
+}
+
+// isProofRequired determines whether a proof is required for the given session's
+// claim based on the current proof module governance parameters.
+// TODO_TECHDEBT: Once the on/off-chain loggers are unified, move this logic to
+// a shared helper used by both off-chain and on-chain routines.
+func (rs *relayerSessionsManager) isProofRequired(
+	ctx context.Context,
+	sessionTree relayer.SessionTree,
+	proofPathSeedBlock client.Block,
+) (isProofRequired bool, err error) {
+	logger := rs.logger.With(
+		"session_id", sessionTree.GetSessionHeader().GetSessionId(),
+		"claim_root", fmt.Sprintf("%x", sessionTree.GetClaimRoot()),
+		"supplier_operator_address", sessionTree.GetSupplierOperatorAddress().String(),
+	)
+
+	// Create the claim object and use its methods to determine if a proof is required.
+	claim := claimFromSessionTree(sessionTree)
+
+	// Get the number of compute units accumulated through the given session.
+	numClaimComputeUnits, err := claim.GetNumComputeUnits()
+	if err != nil {
+		return false, err
+	}
+
+	proofParams, err := rs.proofQueryClient.GetParams(ctx)
+	if err != nil {
+		return false, err
+	}
+
+	logger = logger.With(
+		"num_claim_compute_units", numClaimComputeUnits,
+		"proof_requirement_threshold", proofParams.GetProofRequirementThreshold(),
+	)
+
+	// Require a proof if the claim's compute units meets or exceeds the threshold.
+	if numClaimComputeUnits >= proofParams.GetProofRequirementThreshold() {
+		logger.Info().Msg("compute units is above threshold, claim requires proof")
+
+		return true, nil
+	}
+
+	// Get the hash of the claim to seed the random number generator.
+	var claimHash []byte
+	claimHash, err = claim.GetHash()
+	if err != nil {
+		return false, err
+	}
+
+	// Append the hash of the proofPathSeedBlock to the claim hash to seed the random number generator
+	// to ensure that the proof requirement probability is unknown until the proofPathSeedBlock is observed.
+	// The on-chain claim settlement routine will use the same seed to determine if a proof is required.
+	proofRequirementSeed := append(claimHash, proofPathSeedBlock.Hash()...)
+
+	// Sample a pseudo-random value between 0 and 1 to determine if a proof is required probabilistically.
+	var randFloat float32
+	randFloat, err = poktrand.SeededFloat32(proofRequirementSeed)
+	if err != nil {
+		return false, err
+	}
+
+	logger = logger.With(
+		"claim_hash", fmt.Sprintf("%x", claimHash),
+		"rand_float", randFloat,
+		"proof_request_probability", proofParams.GetProofRequestProbability(),
+	)
+
+	// Require a proof probabilistically based on the proof_request_probability param.
+	// NB: A random value between 0 and 1 will be less than or equal to proof_request_probability
+	// with probability equal to the proof_request_probability.
+	if randFloat <= proofParams.GetProofRequestProbability() {
+		logger.Info().Msg("claim hash seed is below proof request probability, claim requires proof")
+
+		return true, nil
+	}
+
+	logger.Info().Msg("claim does not require proof")
+	return false, nil
+}
+
+// claimFromSessionTree creates a Claim object from the given SessionTree.
+func claimFromSessionTree(sessionTree relayer.SessionTree) prooftypes.Claim {
+	return prooftypes.Claim{
+		SupplierOperatorAddress: sessionTree.GetSupplierOperatorAddress().String(),
+		SessionHeader:           sessionTree.GetSessionHeader(),
+		RootHash:                sessionTree.GetClaimRoot(),
+	}
 }
